Update nvim
This is what it took to install my neovim config on a new system. I have
no idea why it went so wrong. I had to manually install
mason-lspconfig.nvim, which is why it disappeared from the lock file.
For some reason, it was asking for my github password. That might mean
that I had misspelled it, or that the repo was renamed. Idk, but it
works now